refactor(lib:design-system): :truck: Move content page & spec resolvers to design system's content-templates
