+++ apps/humanatlas.io/src/app/app.config.ts
@@ -12,12 +12,12 @@ import { ImageDef } from '@hra-ui/design-system/content-templates/image';
 import { MarkdownDef } from '@hra-ui/design-system/content-templates/markdown';
 import { PageSectionDef } from '@hra-ui/design-system/content-templates/page-section';
 import { VersionedDataTableDef } from '@hra-ui/design-system/content-templates/versioned-data-table';
+import { YouTubePlayerDef } from '@hra-ui/design-system/content-templates/youtube-player';
 import { DataViewerDef } from '@hra-ui/design-system/data-viewer';
 import { PageTableDef } from '@hra-ui/design-system/table';
 import { provideMarkdown } from 'ngx-markdown';
 import { appRoutes } from './app.routes';
 import { ReleaseNotesVersionSelectorDef } from './components/release-notes-version-selector/release-notes-version-selector.definition';
-import { YouTubePlayerDef } from '@hra-ui/design-system/content-templates/youtube-player';
 
 /** Application configuration */
 export const appConfig: ApplicationConfig = {
@@ -31,8 +31,8 @@ export const appConfig: ApplicationConfig = {
       MarkdownDef,
       PageSectionDef,
       PageTableDef,
-      ReleaseNotesVersionSelectorDef,
       ProfileCardDef,
+      ReleaseNotesVersionSelectorDef,
       TextHyperlinkDef,
       VersionedDataTableDef,
       YouTubePlayerDef,
