+++ libs/design-system/content-templates/content-page/src/index.ts
@@ -1,0 +1,2 @@
+export { ContentPageComponent } from './lib/content-page.component';
+export { ContentPageData, ContentPageDataSchema } from './lib/types/content-page.schema';
