+++ libs/design-system/content-templates/content-page/src/lib/content-page.component.scss
