+++ libs/design-system/content-templates/content-page/src/lib/content-page.component.ts
@@ -8,7 +8,7 @@ import { ButtonsModule } from '@hra-ui/design-system/buttons';
 import { PageSectionComponent } from '@hra-ui/design-system/content-templates/page-section';
 import { TableOfContentsLayoutModule } from '@hra-ui/design-system/layouts/table-of-contents';
 import { MarkdownModule } from 'ngx-markdown';
-import { ContentPageData } from '../../schemas/content-page/content-page.schema';
+import { ContentPageData } from './types/content-page.schema';
 
 /**
  * Content Page Component
@@ -19,11 +19,11 @@ import { ContentPageData } from '../../schemas/content-page/content-page.schema'
     HraCommonModule,
     ButtonsModule,
     ContentTemplateOutletDirective,
+    MarkdownModule,
     MatIconModule,
     PageSectionComponent,
     RouterModule,
     TableOfContentsLayoutModule,
-    MarkdownModule,
   ],
   templateUrl: './content-page.component.html',
   styleUrl: './content-page.component.scss',
