+++ libs/design-system/content-templates/resolvers/src/index.ts
@@ -1,0 +1,1 @@
+export { createJsonSpecResolver, createYamlSpecResolver } from './lib/spec.resolver';
