+++ libs/design-system/content-templates/resolvers/src/lib/spec.resolver.ts
